--- a/fms/models/llama.py
+++ b/fms/models/llama.py
@@ -126,7 +126,7 @@ class LLaMABlock(nn.Module):
         residual = x
         x = self.ln(x)
         x = self.attn(
-            qkv=x,
+            q=x,
             mask=mask,
             position_ids=position_ids,
             attn_algorithm=attn_algorithm,