--- a/fms/modules/attention.py
+++ b/fms/modules/attention.py
@@ -133,7 +133,7 @@ class MultiHeadAttention(nn.Module):
             in past_key_value_state
         """
 
-        if k is None and v is None:
+        if (k is None and v is None) or (k is q and v is q):
             qkv = q
         elif k is not None and v is not None:
             # Note: for encoder/decoder models with cross attn, this line may need to be changed as kv will be fused and