Many models implementations now have moved to fusing the query/key/value in attention. This has been held off as we did not have a clean way of ingesting to/from fused/unfused weights, but now that we have the safetensors implementation with FusableWeightsMissingError, this is viable. This is one of the optimizations that we found benefitted performance and compile greatly.

Note: Because weight names have been changed from query, key, and value in attention to qkv_fused, expectation tests needed to be re-run

Benchmark Results (~1ms improvement):

Main:

```
Results for batch size 1, sequence length 500, new tokens generated 256
Uncompiled results:
==========
Single token generation
- with use_cache=True
	27.47 ms per token
End-to-end sequence generation
- with use_cache=True
	27.31 ms per token

Results for batch size 2, sequence length 500, new tokens generated 256
Uncompiled results:
==========
Single token generation
- with use_cache=True
	27.61 ms per token
End-to-end sequence generation
- with use_cache=True
	27.55 ms per token

Results for batch size 8, sequence length 500, new tokens generated 256
Uncompiled results:
==========
Single token generation
- with use_cache=True
	27.41 ms per token
End-to-end sequence generation
- with use_cache=True
	28.27 ms per token
```

fused_attn

```
Results for batch size 1, sequence length 500, new tokens generated 256
Uncompiled results:
==========
Single token generation
- with use_cache=True
	26.92 ms per token
End-to-end sequence generation
- with use_cache=True
	26.75 ms per token

Results for batch size 2, sequence length 500, new tokens generated 256
Uncompiled results:
==========
Single token generation
- with use_cache=True
	26.46 ms per token
End-to-end sequence generation
- with use_cache=True
	26.47 ms per token

Results for batch size 8, sequence length 500, new tokens generated 256
Uncompiled results:
==========
Single token generation
- with use_cache=True
	26.60 ms per token
End-to-end sequence generation
- with use_cache=True
	27.71 ms per token
```