@nairbv 2024-02-13
how much of a performance improvement do we see from this fusion? I think we should add the benchmark output for this change in the PR desc